+++ src/datasources/alarms/constants/DefaultQueries.constants.ts
@@ -1,4 +1,5 @@
 import { AlarmsCountQuery } from '../types/AlarmsCount.types';
+import { AlarmsTrendQuery } from '../types/AlarmsTrend.types';
 import { AlarmsProperties, ListAlarmsQuery } from '../types/ListAlarms.types';
 import { AlarmsVariableQuery, QueryType } from '../types/types';
 import { DEFAULT_QUERY_EDITOR_DESCENDING, DEFAULT_QUERY_EDITOR_TAKE, DEFAULT_QUERY_EDITOR_TRANSITION_INCLUSION_OPTION } from './AlarmsQueryEditor.constants';
@@ -19,7 +20,13 @@ export const defaultListAlarmsQuery: Omit<ListAlarmsQuery, 'refId'> = {
     AlarmsProperties.state,
     AlarmsProperties.workspace,
   ],
+  take: DEFAULT_QUERY_EDITOR_TAKE,
+  descending: DEFAULT_QUERY_EDITOR_DESCENDING,
   transitionInclusionOption: DEFAULT_QUERY_EDITOR_TRANSITION_INCLUSION_OPTION,
+};
+
+export const defaultAlarmsTrendQuery: Omit<AlarmsTrendQuery, 'refId'> = {
+  filter: '',
 };
 
 export const defaultListAlarmsVariableQuery: Omit<AlarmsVariableQuery, 'refId'> = {