+++ src/datasources/alarms/types/ListAlarms.types.ts
@@ -3,6 +3,8 @@ import { AlarmsQuery, TransitionInclusionOption } from './types';
 export interface ListAlarmsQuery extends AlarmsQuery {
   filter?: string;
   properties?: AlarmsProperties[];
+  descending?: boolean;
+  take?: number;
   transitionInclusionOption?: TransitionInclusionOption;
 }